src/bot/line/filter.ts
import { createFilter } from './handler';
import { getPostBackText, isGroupEvent, isPostBackEvent, isSingleEvent, isTextMessage } from './types';
import { LineBotErrorMessage } from './errors';

export const Group = createFilter(event => (isGroupEvent(event) ? event : false));
export const MemberJoin = createFilter(event => event.type === 'memberJoined');
export const MemberLeft = createFilter(event => event.type === 'memberLeft');
export const LeaveGroup = createFilter(event => event.type === 'leave');

export const Single = createFilter(isSingleEvent);

export const UserId = ({ warning = true } = {}) =>
  createFilter(event => {
    const userId = event.source.userId;
    if (userId) return userId;
    if (warning) throw new LineBotErrorMessage('CANNOT_GET_USER_ID');
  });

export const TextEqual = (options: string | string[], { shouldReturn = true, postbackOnly = false } = {}) => {
  const validate =
    typeof options === 'string'
      ? (payload: string) => options === payload
      : (payload: string) => options.includes(payload);

  return createFilter(event => {
    let text = '';

    if (isPostBackEvent(event)) {
      text = getPostBackText(event);
    } else if (postbackOnly) return;

    if (isTextMessage(event)) {
      text = event.message.text;
    }

    if (!!text && shouldReturn && validate(text)) return text;
  });
};

export const TextMatch = (regex: RegExp, { postbackOnly = false } = {}) => {
  return createFilter(event => {
    let text = '';

    if (isPostBackEvent(event)) {
      text = getPostBackText(event);
    } else if (postbackOnly) return;

    if (isTextMessage(event)) {
      text = event.message.text;
    }

    return !!text && text.match(regex);
  });
};
